+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Bug #18: Update assertion message for integer setting in `StrokeWidthProvider` class (@terabytesoftw)
 - Enh #19: Add `HasStrokeDashArray` trait and corresponding tests for managing SVG `stroke-dasharray` attribute (@terabytesoftw)
 - Bug #20: Update `HasStrokeDashArray` trait and tests to support float values for `stroke-dasharray` attribute (@terabytesoftw)
+- Enh #21: Add `HasOpacity` trait and corresponding tests for managing SVG `opacity` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024